Case rename mbcup selectors (2/2)
